IDE-1508.
Modify the final local var analysis to perform only the following:
1. verify there can be no code path between a final var declaration and a reference to it where the var is not initialized
2. verify there can be no code path between a final var declaration and an assignment to it where the var is initialized
Note we already perform these two, so the change is a small matter of removing a third part where we verify the final local var is initialized in its use-scope.  This proves to be more confusing than it's worth -- the local var isn't like a field so much, so we don't really need to verify that it is covered re initialization.