Adding "Known Break" feature to BaseVerifyErrantTest
A test class can have lines of code with makers like "//## KB(JIRA-123) you can write an optional message here"
This maker stands for a known break at that line and it is described in the jira JIRA-123.
These markers make BaseVerifyErrantTest to fail.
This behaviour is switched on and off by the system property gw.tests.skip.knownbreak.